+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/MapboxMap.kt
@@ -7,6 +7,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.viewinterop.AndroidView
 import com.mapbox.maps.*
+import com.mapbox.maps.extension.compose.animation.viewport.MapViewportState
+import com.mapbox.maps.extension.compose.animation.viewport.rememberMapViewportState
 import com.mapbox.maps.extension.compose.internal.MapApplier
 import com.mapbox.maps.extension.compose.internal.MapPreviewPlaceHolder
 import com.mapbox.maps.extension.compose.internal.MapboxMapComposeNode
@@ -22,74 +24,44 @@ import kotlinx.coroutines.awaitCancellation
 
 /**
  * Entry point for adding a Mapbox Map instance to the Jetpack Compose UI.
+ *
+ * @param modifier Modifier to be applied to the Mapbox map.
+ * @param mapInitOptionsFactory Defines the initialisation configurations factory for a [MapboxMap]. It can only be set once and not mutable after the initialisation. Mutating the [MapInitOptions] during recomposition will result in a [IllegalStateException].
+ * @param attributionSettings Settings for showing the attribution icon on the map.
+ * @param compassSettings Settings for showing the compass icon on the map.
+ * @param gesturesSettings Gesture configuration allows to control the user touch interaction.
+ * @param locationComponentSettings Settings for showing a location puck on the map.
+ * @param logoSettings Settings for showing the Mapbox logo on the map.
+ * @param scaleBarSettings Settings for showing the scale bar on the map.
+ * @param mapViewportState A state object that can be hoisted to control and observe the map's camera state. A [MapViewportState] may only be used by a single [MapboxMap] composable at a time as it reflects instance state for a single view of a map.
+ * @param onMapClickListener Callback to be invoked when the user clicks on the map view.
+ * @param onMapLongClickListener Callback to be invoked when the user long clicks on the map view.
+ * @param content The content of the map.
  */
 @Composable
 @MapboxExperimental
 public fun MapboxMap(
-  /**
-   * Modifier to be applied to the Mapbox map.
-   */
   modifier: Modifier = Modifier,
-  /**
-   * Defines the initialisation configurations factory for a [MapboxMap].
-   *
-   * It can only be set once and not mutable after the initialisation. Mutating the [MapInitOptions]
-   * during recomposition will result in a [IllegalStateException].
-   */
   mapInitOptionsFactory: (Context) -> MapInitOptions = { context -> MapInitOptions(context) },
-  /**
-   * Settings for showing the attribution icon on the map.
-   */
   attributionSettings: AttributionSettings = DefaultSettingsProvider.defaultAttributionSettings(
     LocalContext.current.applicationContext
   ),
-  /**
-   * Settings for showing the compass icon on the map.
-   */
   compassSettings: CompassSettings = DefaultSettingsProvider.defaultCompassSettings(
     LocalContext.current.applicationContext
   ),
-  /**
-   * Gesture configuration allows to control the user touch interaction.
-   */
   gesturesSettings: GesturesSettings = DefaultSettingsProvider.defaultGesturesSettings,
-  /**
-   * Settings for showing a location puck on the map.
-   */
   locationComponentSettings: LocationComponentSettings = DefaultSettingsProvider.defaultLocationComponentSettings(
     LocalContext.current.applicationContext,
   ),
-  /**
-   * Settings for showing the Mapbox logo on the map.
-   */
   logoSettings: LogoSettings = DefaultSettingsProvider.defaultLogoSettings(
     LocalContext.current.applicationContext
   ),
-  /**
-   * Settings for showing the scale bar on the map.
-   */
   scaleBarSettings: ScaleBarSettings = DefaultSettingsProvider.defaultScaleBarSettings(
     LocalContext.current.applicationContext
   ),
-  /**
-   * Set the camera of the map.
-   */
-  cameraOptions: CameraOptions = CameraOptions.Builder().build(),
-  /**
-   * Lambda to be invoked when camera changes.
-   */
-  onCameraStateChange: (CameraState) -> Unit = DefaultSettingsProvider.defaultOnCameraStateChange,
-  /**
-   * Callback to be invoked when the user clicks on the map view.
-   */
+  mapViewportState: MapViewportState = rememberMapViewportState(),
   onMapClickListener: OnMapClickListener = DefaultSettingsProvider.defaultOnClickListener,
-  /**
-   * Callback to be invoked when the user long clicks on the map view.
-   */
   onMapLongClickListener: OnMapLongClickListener = DefaultSettingsProvider.defaultOnLongClickListener,
-  /**
-   * The content of the map.
-   */
   content: (@Composable @MapboxMapComposable MapboxMapScope.() -> Unit)? = null
 ) {
   // display placeholder when in preview mode.
@@ -119,8 +91,7 @@ public fun MapboxMap(
   val currentLocationComponentSettings by rememberUpdatedState(locationComponentSettings)
   val currentLogoSettings by rememberUpdatedState(logoSettings)
   val currentScaleBarSettings by rememberUpdatedState(scaleBarSettings)
-  val currentCameraOptions by rememberUpdatedState(cameraOptions)
-  val currentOnCameraStateChange by rememberUpdatedState(onCameraStateChange)
+  val currentMapViewportState by rememberUpdatedState(mapViewportState)
   val currentOnMapClickListener by rememberUpdatedState(onMapClickListener)
   val currentOnMapLongClickListener by rememberUpdatedState(onMapLongClickListener)
   val currentContent by rememberUpdatedState(content)
@@ -139,8 +110,7 @@ public fun MapboxMap(
             currentLocationComponentSettings,
             currentLogoSettings,
             currentScaleBarSettings,
-            currentCameraOptions,
-            currentOnCameraStateChange,
+            currentMapViewportState,
             currentOnMapClickListener,
             currentOnMapLongClickListener
           )